Bot working AI and some bugs remains

+++ Utils/Vector2.cs
@@ -8,9 +8,10 @@ namespace LandFightBotReborn.Utils
 {
     public class Vector2
     {
-        public int x;
-        public int y;
-        public Vector2(int x, int y)
+        public float x;
+        public float y;
+        
+        public Vector2(float x, float y)
         {
             this.x = x;
             this.y = y;
